+++ components/persistor/test/persist_newapi_tests.js
@@ -426,7 +426,8 @@ describe('persistor transaction checks', function () {
                     innerTxn.postSave = function(txn, _logger, changes) {
                         expect(Object.keys(changes)).to.contain('Employee');
                         expect(Object.keys(changes.Employee[0])).to.contain('primaryKey');
-                        expect(changes.Employee[0].properties[0].name).to.equal('dob');
+                        expect(changes.Employee[0].properties[0].name).to.equal('homeAddress');
+                        expect(changes.Employee[0].properties[1].name).to.equal('dob');
                         var empNew = new Employee();
                         empNew.setDirty(txn);
                     };